requirements.txt
asn1crypto==0.22.0
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
bumpversion==0.5.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
codacy-coverage==1.3.6
configparser==3.5.0
coverage==4.4.1
cryptography==2.0.3
decorator==4.1.2
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.8.18
django-bootstrap-form==3.3
django-datetime-widget==0.9.3
django-debug-toolbar==1.8
django-registration-redux==1.7
eight==0.4.2
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
jdcal==1.3
Jinja2==2.9.6
lml==0.0.1
lxml==3.8.0
markdown2==2.3.4
MarkupSafe==1.0
mccabe==0.6.1
MySQL-python==1.2.5
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.2.1
pipdeptree==0.10.1
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pyasn1==0.3.2
pycodestyle==2.3.1
pycparser==2.18
pyexcel==0.5.3
pyexcel-io==0.4.4
pyexcel-ods==0.4.0
pyexcel-xlsx==0.4.1
pyflakes==1.6.0
Pygments==2.2.0
PyMySQL==0.7.11
pyOpenSSL==17.2.0
python-memcached==1.58
python-vote-full==1.0
pytz==2017.2
requests==2.18.4
scandir==1.5
signxml==2.4.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
suds==0.4
texttable==0.9.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
